[Refactor] Move FSNotify trigger into a separate package
In preparation for #5140,
1. Move fsNotifyTrigger into its own package.
2. Add a function `Ignore` to configure the fsNotifyWatcher to ignore
   - files belonging to predefined dirs like .idea .vscode & .git
   - hidden file like .somefile.swp

*Output change**
None.

*Future Work*
1. Configure the `Ignore` to ignore files mentioned above
